Improve fuzzy matching of programs.
Sorts by Jaccard index of case insensitive comparison of the user's
query and the entry title. Match an entry if it either entirely contains
the user's query or if it matches a regex created by putting wildcards
between every character of the query.

For example, "Google Chrome" matches "gc", and "Sublime Text 3" matches
"st3". It's similar to what Sublime Text 3's "Go to Anything..." feature
does, but slightly less fancy.