Refactor updating results.
It's a little bit cleaner to have the logic for whether to show the
footer inside updateResults. Also, updateResults is a way better name.
Perhaps consider moving all of the DOM update functions into one module.
Also, we definitely don't need to hide an empty div; that was just
silly to begin with.