Add server component.
Fixes #3. Sort of makes it launch faster, but not really anything
substantial until we move the logic for reading files, etc. over. The
main problem I can see with loading .desktop files in the server is that
they won't really refresh unless you watch the directories for new
files. So if we decide to stick with this version, we should carefully
consider that aspect of it.